requirements.txt
pandas
openai
psycopg[binary]
python-dotenv
instructor
anthropic